fix warnings for dtype with vsc-hvdc

+++ pandapower/pf/makeYbus_facts.py
@@ -187,7 +187,7 @@ def makeYbus_ssc_vsc(Ybus, internal_y_pu, fb, tb, controllable):
     return Ybus_not_controllable, Ybus_controllable, Ybus_not_controllable + Ybus_controllable
 
 
-def make_Ybus_facts(from_bus, to_bus, y_pu, n):
+def make_Ybus_facts(from_bus, to_bus, y_pu, n, dtype=np.complex128):
     """
     Construct the bus admittance matrix with an added FACTS device for a power grid.
 
@@ -214,7 +214,7 @@ def make_Ybus_facts(from_bus, to_bus, y_pu, n):
     data = np.concatenate([y_pu, y_pu, -y_pu, -y_pu])
 
     # Create and return the Ybus matrix using the compressed sparse row format
-    Ybus_facts = csr_matrix((data, (row_indices, col_indices)), shape=(n, n), dtype=np.complex128)
+    Ybus_facts = csr_matrix((data, (row_indices, col_indices)), shape=(n, n), dtype=dtype)
     return Ybus_facts
